requirements.txt
celery==5.2.7
charset-normalizer==3.2.0
crispy-bootstrap5==0.7
Django==4.1.4
django-crispy-forms==1.14.0
django-emoji-picker==0.0.6
djangorestframework==3.14.0
gevent==22.10.2
Pillow
psycopg2-binary